coldata: move RandomVec into coldata
Release note: None
