+++ pkg/sql/colexec/projection_ops_test.go
@@ -217,8 +217,8 @@ func TestRandomComparisons(t *testing.T) {
 		lVec := b.ColVec(0)
 		rVec := b.ColVec(1)
 		ret := b.ColVec(2)
-		RandomVec(rng, typ, bytesFixedLength, lVec, numTuples, 0)
-		RandomVec(rng, typ, bytesFixedLength, rVec, numTuples, 0)
+		coldata.RandomVec(rng, typ, bytesFixedLength, lVec, numTuples, 0)
+		coldata.RandomVec(rng, typ, bytesFixedLength, rVec, numTuples, 0)
 		for i := range lDatums {
 			lDatums[i] = PhysicalTypeColElemToDatum(lVec, uint16(i), da, ct)
 			rDatums[i] = PhysicalTypeColElemToDatum(rVec, uint16(i), da, ct)
